+++ docs/changes.rst
@@ -4,7 +4,7 @@ Changlog
 Version 0.5.0
 -------------
 
-To be released.
+Released on July 24, 2017.
 
 - Added ``recurse`` option to :class:`~settei.base.config_object_property`.
   If it's :const:`True` nested tables are also evaluated.  :const:`False` by